parking/pyproject.toml
[tool.poetry]
name = "parking"
version = "0.1.0"
description = ""
authors = ["Aayam <aayam.maharjan3@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.14"
celery = "5.4.0"
redis = "5.0.7"
sudo = "1.0.0"
django-celery-results = "2.5.1"
django-celery-beat = "2.6.0"
django-decouple = "2.1"
django-sphinxdoc = "2.0.0"
pytest = "8.3.2"
pytest-django = "4.8.0"
factory-boy = "3.3.0"
django-import-export = "4.1.1"
pandas = "2.2.2"
django-environ = "0.9.0"
django-widget-tweaks = "1.5.0"
sweetify = "2.3.1"
djangorestframework = "3.15.2"
amqp = "5.2.0"
asgiref = "3.8.1"
billiard = "4.2.0"
cfgv = "3.4.0"
click = "8.1.7"
click-didyoumean = "0.3.1"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
cron-descriptor = "1.4.3"
diff-match-patch = "20230430"
distlib = "0.3.8"
django-extensions = "3.2.3"
django-haystack = "3.3.0"
django-restframework = "0.0.1"
django-timezone-field = "7.0"
faker = "26.0.0"
filelock = "3.15.4"
identify = "2.6.0"
iniconfig = "2.0.0"
kombu = "5.3.7"
nodeenv = "1.9.1"
numpy = "2.0.1"
packaging = "24.1"
platformdirs = "4.2.2"
pluggy = "1.5.0"
pre-commit = "3.8.0"
prompt-toolkit = "3.0.47"
python-crontab = "3.2.0"
python-dateutil = "2.9.0.post0"
pytz = "2024.1"
pyyaml = "6.0.1"
setuptools = "72.0.0"
six = "1.16.0"
sqlparse = "0.5.1"
tablib = "3.5.0"
tzdata = "2024.1"
vine = "5.1.0"
virtualenv = "20.26.3"
wcwidth = "0.2.13"
wheel = "0.43.0"
django-filter = "^24.2"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"
django-translations = "^1.3.0"
djangorestframework-simplejwt = "^5.3.1"
rosetta = "^0.3"
django-modeltranslation = "^0.19.5"
django-cors-headers = "^4.4.0"
scipy = "^1.14.0"
nltk = "^3.8.1"
django-rosetta = "^0.10.0"
whitenoise = "^6.7.0"
translation = "^1.0.5"
environ = "^1.0"
pymupdf = "^1.24.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
django-extensions = "^3.2.3"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"